requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.5
django-etc==1.4.0
django-hitcount==1.3.5
djangorestframework==3.14.0
Faker==16.1.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.15